Support @SerialId on enum members in protobuf encoding
