+++ runtime/commonMain/src/kotlinx/serialization/protobuf/ProtoTypes.kt
@@ -1,17 +1,5 @@
 /*
- * Copyright 2018 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright 2017-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package kotlinx.serialization.protobuf
@@ -28,8 +16,8 @@ public annotation class ProtoType(val type: ProtoNumberType)
 
 internal typealias ProtoDesc = Pair<Int, ProtoNumberType>
 
-internal fun extractParameters(desc: SerialDescriptor, index: Int): ProtoDesc {
-    val idx = getSerialId(desc, index) ?: index + 1
+internal fun extractParameters(desc: SerialDescriptor, index: Int, zeroBasedDefault: Boolean = false): ProtoDesc {
+    val idx = getSerialId(desc, index) ?: (if (zeroBasedDefault) index else index + 1)
     val format = desc.findAnnotation<ProtoType>(index)?.type
             ?: ProtoNumberType.DEFAULT
     return idx to format
